feat: 댓글 미리 보기 서비스 메서드 구현

+++ src/main/java/sleepy/mollu/server/content/comment/controller/dto/SearchCommentPreviewResponse.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import java.time.LocalDateTime;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record SearchCommentPreviewResponse(int commentNum, CommentResponse comment) {
+public record SearchCommentPreviewResponse(Long commentNum, CommentResponse comment) {
 
     public record CommentResponse(String commentId, String comment, String molluId, String userName,
                                   String profileSource, LocalDateTime createdAt) {
